+++ Assets/VRM10/Runtime/Components/Expression/IExpressionAccumulator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fdd7cdf6d59574348bb7336272ef5edf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
